requirements.txt
absl-py==2.2.2
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.12.0
arrow==1.4.0
asttokens==3.0.0
astunparse==1.6.3
attrs==23.2.0
beautifulsoup4==4.12.3
bitarray==3.7.1
bleach==6.2.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
ckzg==2.1.5
click==8.3.1
cloudscraper==1.2.71
codecarbon==3.2.0
colorama==0.4.6
Columnar==1.4.1
contourpy==1.2.1
cryptography==46.0.3
cycler==0.12.1
cytoolz==1.0.1
DateTime==5.5
decorator==5.2.1
discord.py==2.3.2
et_xmlfile==2.0.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
executing==2.2.1
fake-useragent==1.4.0
fawkes==1.0.3
fief-client==0.20.0
filelock==3.20.1
flatbuffers==25.2.10
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2025.12.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h11==0.16.0
h5py==3.13.0
HashTable==0.1
hexbytes==1.3.1
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.27.2
huggingface-hub==0.36.0
hyperliquid-python-sdk==0.19.0
idna==3.11
imageio==2.37.0
iniconfig==2.1.0
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jwcrypto==1.5.6
keras==3.10.0
kiwisolver==1.4.5
libclang==18.1.1
lxml==5.3.0
lz4==4.4.4
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.3
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.1
mtcnn==1.0.0
multidict==6.0.5
namex==0.0.9
networkx==3.6.1
numpy==1.26.4
nvidia-ml-py==13.580.82
opencv-python==4.12.0.88
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.15.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.1
parsimonious==0.10.0
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
playsound==1.3.0
pluggy==1.6.0
poly_eip712_structs==0.0.1
prometheus_client==0.23.1
prompt_toolkit==3.0.52
protobuf==5.29.4
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py_clob_client==0.25.0
py_order_utils==0.3.2
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.9
pydantic_core==2.33.2
pygame==2.5.2
Pygments==2.19.1
pyparsing==3.1.1
PySocks==1.7.1
pytest==8.4.2
python-dateutil==2.8.2
python-dotenv==1.1.1
pytz==2024.1
PyYAML==6.0.3
questionary==2.1.1
RapidFuzz==3.14.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.0.0
rlp==4.1.0
safetensors==0.7.0
scipy==1.15.2
selenium==4.25.0
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.2
soupsieve==2.5
stack-data==0.6.3
sterra==2.2.7
string-color==1.3.0
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
timm==1.0.22
tokenizers==0.22.1
toolz==1.0.0
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.3
trio==0.27.0
trio-websocket==0.11.1
typer==0.20.0
typer-slim==0.21.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.1
urllib3==2.6.2
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.42.0
wrapt==1.17.2
wsproto==1.2.0
yarl==1.9.4
yaspin==3.3.0
zope.interface==8.0.1
